Add checks and errors in mnDQM for bad time range
mnDQM requires that the time range of the GCM cover the full range of
the calibrated historical period and extend into the future. If this
is not the case (i.e. either the GCM is too small or the calibration
period is misconfigured), this commit adds an error message explaining
the problem.